# Improve performance of SqlString.escape()