--- a/lib/protocol/SqlString.js
+++ b/lib/protocol/SqlString.js
@@ -26,40 +26,21 @@ SqlString.escape = function escape(val, stringifyObjects, timeZone) {
   switch (typeof val) {
     case 'boolean': return (val) ? 'true' : 'false';
     case 'number': return val+'';
+    case 'object':
+      if (val instanceof Date) {
+        val = SqlString.dateToString(val, timeZone || 'local');
+      } else if (Array.isArray(val)) {
+        return SqlString.arrayToList(val, timeZone);
+      } else if (Buffer.isBuffer(val)) {
+        return SqlString.bufferToString(val);
+      } else if (stringifyObjects) {
+        val = val.toString();
+      } else {
+        return SqlString.objectToValues(val, timeZone);
+      }
   }
 
-  if (val instanceof Date) {
-    val = SqlString.dateToString(val, timeZone || 'local');
-  }
-
-  if (Buffer.isBuffer(val)) {
-    return SqlString.bufferToString(val);
-  }
-
-  if (Array.isArray(val)) {
-    return SqlString.arrayToList(val, timeZone);
-  }
-
-  if (typeof val === 'object') {
-    if (stringifyObjects) {
-      val = val.toString();
-    } else {
-      return SqlString.objectToValues(val, timeZone);
-    }
-  }
-
-  val = val.replace(/[\0\n\r\b\t\\\'\"\x1a]/g, function(s) {
-    switch(s) {
-      case "\0": return "\\0";
-      case "\n": return "\\n";
-      case "\r": return "\\r";
-      case "\b": return "\\b";
-      case "\t": return "\\t";
-      case "\x1a": return "\\Z";
-      default: return "\\"+s;
-    }
-  });
-  return "'"+val+"'";
+  return escapeString(val);
 };
 
 SqlString.arrayToList = function arrayToList(array, timeZone) {
@@ -155,6 +136,37 @@ SqlString.objectToValues = function objectToValues(object, timeZone) {
 
   return sql;
 };
+
+var charsRegex = /[\0\b\t\n\r\x1a\"\'\\]/g;
+var charsMap = {
+  '\0': '\\0',
+  '\b': '\\b',
+  '\t': '\\t',
+  '\n': '\\n',
+  '\r': '\\r',
+  '\x1a': '\\Z',
+  '"': '\\"',
+  '\'': '\\\'',
+  '\\': '\\\\'
+};
+function escapeString(val) {
+  var chunkIndex = charsRegex.lastIndex = 0;
+  var escapedVal = '';
+  var match;
+
+  while ((match = charsRegex.exec(val))) {
+    escapedVal += val.slice(chunkIndex, match.index) + charsMap[match[0]];
+    chunkIndex = charsRegex.lastIndex;
+  }
+
+  if (chunkIndex === 0) { // Nothing was escaped
+    return "'" + val + "'";
+  }
+  if (chunkIndex < val.length) {
+    return "'" + escapedVal + val.slice(chunkIndex) + "'";
+  }
+  return "'" + escapedVal + "'";
+}
 
 function zeroPad(number, length) {
   number = number.toString();