Strategies include:
+ Manually escaping strings rather than using a string replace function with a regex
+ Only calling `typeof` on the input value once
+ Not performing extra type checks if the input type has already been determined

#### Benchmark

See [this gist](https://gist.github.com/nwoltman/98dcc294b77da3bedad9bb89e54ab2fa) for the code.

| Value | Before (ops/sec) | This PR (ops/sec) | Speedup (`PR/before`) |
|---------|----------------------|------------------------|-------------|
| `'lastname'` | 4,916,714 | 13,443,318 | 2.73 |
| `'firstname123@gmail.com'` | 4,843,172 | 13,258,714 | 2.74 |
| `'to "escape"'` | 1,846,340 | 3,510,579 | 1.90 |
| `'Lorem ipsum...'` | 1,724,015 | 2,212,050 | 1.28 |
| `'Lorem ipsum...'` with `'\n'` | 845,988 | 1,346,465 | 1.59 |
| JSON | 486,268 | 493,629 | 1.02 |