@dougwilson 2016-04-27
Nice! Code certainly looks weird with the slicing, but I guess if it's faster... :) And thanks a lot for providing that gist with the benchmark code! That would allow us to play with this a bit, for example, to test how it changed very long strings instead of just short ones (only because I have in the past seem optimizations that help one case, but really hurt another case), to have a better overall idea. Otherwise, I think it should be merged soon!