@nwoltman 2016-04-28
Thanks! And that's a good idea. I updated the benchmark to include longer strings and it looks like the longer the string gets, the slower my implementation becomes. Would you agree that I should use the original code as a fallback for long strings?