@dougwilson 2016-04-28
Perhaps. Are you saying that the longer the string becomes, the more this change actually becomes a de-optimization compared to the original implementation?

I have tried a few times to do "magic numbers" where an implementation will flip to a different one (where one is faster for the short case and one is faster for the long case), but it can sometimes be hard to really know at which number to make the flip (the number may very well different depending on the v8 engine version).

I guess one question I would ask myself is "is this the most optimal implementation, or was it coincidentally optimal for some small case?" What I mean by that is I have seen, when only looking at the small case, that sometimes there are different implementations that can get both short and long cases that were just not immediately obvious when only focusing on the short case.

Anyway, I think that, at the very least, probably adding a magic implementation switch would be good, but also if you want to think some more on the optimization, feel free :)