@nwoltman 2016-04-28
Yes, that is what I meant to say.

Unfortunately the only optimizations left that I know of are micro-optimizations, and their small performance benefits are not worth making the code less readable. I know my current implementation cannot be optimized further because of the following: The case where the `.replace()` call is much faster is when the input is a long string that doesn't have any characters to be escaped, so a majority of my code never get's used. In that case, even if I take out all of the unused code, the `.replace()` method is still about 4x faster (against "Lorem Ipsum"). For reference, here's what my code would look like without the unused code:

```js
function escapeString(val) {
  for (var i = 0; i < val.length; i++) {
    switch (val.charCodeAt(i)) {
    }
  }
  return "'" + val + "'";
}
```

So in that sense, it cannot be optimized more. However, this gives us a better idea of why my method is faster when either a) the string is short, or b) the string has many characters that need to be escaped. I'm assuming my method is faster because it is 100% client JS where as with `.replace()` it has to go back and forth between the client code/provided callback and the V8 internals. With the long, plain string, `.replace()` wins because it only goes into the V8 internals once, and the V8 code is faster than just JS code.

Now there are two things that can still be done. The first is trying to find the magic number. The second is putting this somewhere:

```js
if (!/[\0\b\t\n\r\x1a\"\'\\]/.test(val)) {
  return "'" + val + "'";
}
```

Even though that reduces performance a little for short strings and strings that contain characters that need to be escaped, it greatly improves performance for longer strings that do not contain characters that need to be escaped.

One option is to optimize more for the common cases without de-optimizing the rarer cases. Do you have any suggestions on what the common cases might be? From what I've seen, most strings don't have any of those characters that need escaping.