@dougwilson 2016-04-28
Awesome, I love this research :) 

> Do you have any suggestions on what the common cases might be?

That is definitely a good question, and I'm not too sure off-hand. Just thinking about some of my MySQL usages that are out there, most of my calls to `.escape` is just with a number primitive, with the second most being strings. From the string perspective, I image the vast majority do not conain any characters that need to be escaped, even long strings. If they do have characters that need to be escaped, it would be due to the single or the double quote characters. Certainly the cases where I am using long strings are typically going to be storing some kind of serialization, either HTML or JSON, both of which typically contain many double quote characters (and single quotes for HTML).

So from just thinking out loud to myself, perhaps a more common case for long strings would be (1) nothing is being escaped at all (perhaps it's some kind of simple CSV?) (2) there are lots of escaped characters within the string. I wonder (2) fares over all.

Also, on the note of optimizations, I noticed that in this module, that giant regular expression is a literal within the function. I feel like I've heard that doing that is not great for v8, and that hoisting the regular expression out and into a global helps a lot. I wonder how much that would help here. The function in the `.replace` could also potentionally be hoisted out since it doesn't close over any variable.

Another thing is that I usually find that regular expressions are very fast to execute in the v8 engine. I wonder if there is a way to replace that `for` loop with a regular expression that finds the next character to escape, rather than looking up the char code for each character? Perhaps that's what's eating up all the time in the new function. I'm thinking something along the lines of:
```js
var chars = /[\0\b\t\n\r\x1a\"\'\\]/g;
function escapeString(val) {
  chars.lastIndex = 0;
  var match;
  while((match = chars.exec(val))) {
    switch (val.charCodeAt(match.index)) {
    }
  }
  return "'" + val + "'";
}
```
That is probably pretty cose to what `.replace` does internally, though not sure if it's faster or not :)